+++ app/src/main/java/com/cook/easypan/easypan/presentation/navigation/HomeNavGraph.kt
@@ -1,36 +1,23 @@
 package com.cook.easypan.easypan.presentation.navigation
 
-import ProfileAction
 import android.content.ContentValues.TAG
 import android.util.Log
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.RowScope
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.NavigationBar
-import androidx.compose.material3.NavigationBarItem
-import androidx.compose.material3.NavigationBarItemColors
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavBackStackEntry
 import androidx.navigation.NavController
-import androidx.navigation.NavDestination
-import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import com.cook.easypan.easypan.data.auth.GoogleAuthUiClient
 import com.cook.easypan.easypan.presentation.SelectedRecipeViewModel
 import com.cook.easypan.easypan.presentation.favorite.FavoriteRoot
 import com.cook.easypan.easypan.presentation.favorite.FavoriteViewModel
@@ -51,7 +38,6 @@ import org.koin.androidx.compose.koinViewModel
 
 @Composable
 fun Home(
-    googleAuthUiClient: GoogleAuthUiClient,
     navController: NavHostController = rememberNavController(),
     onSignOut: () -> Unit
 ) {
@@ -66,7 +52,6 @@ fun Home(
         ) {
             HomeNavGraph(
                 navController = navController,
-                googleAuthUiClient = googleAuthUiClient,
                 onSignOut = onSignOut
             )
         }
@@ -74,34 +59,8 @@ fun Home(
 }
 
 @Composable
-fun BottomNavigationBar(navController: NavHostController) {
-    val navBackStackEntry by navController.currentBackStackEntryAsState()
-    val currentDestination = navBackStackEntry?.destination
-
-    val screens = listOf(
-        BottomBarScreen.Home,
-        //BottomBarScreen.Favorite,
-        BottomBarScreen.Profile
-    )
-    val bottomBarDestination =
-        screens.any { it.route::class.qualifiedName == currentDestination?.route }
-    if (bottomBarDestination) {
-        NavigationBar {
-            screens.forEach { screen ->
-                AddItem(
-                    screen = screen,
-                    currentDestination = currentDestination,
-                    navController = navController
-                )
-            }
-        }
-    }
-}
-
-@Composable
 fun HomeNavGraph(
     navController: NavHostController,
-    googleAuthUiClient: GoogleAuthUiClient,
     onSignOut: () -> Unit
 ) {
     NavHost(
@@ -162,8 +121,11 @@ fun HomeNavGraph(
             RecipeStepRoot(
                 viewModel = viewModel,
                 onFinishClick = {
-                    navController.popBackStack()
-                    navController.navigate(Route.RecipeFinish)
+                    navController.navigate(Route.RecipeFinish) {
+                        popUpTo(Route.RecipeDetail) {
+                            inclusive = true
+                        }
+                    }
                 },
                 onCancelClick = {
                     navController.navigateUp()
@@ -183,19 +145,9 @@ fun HomeNavGraph(
 
         composable<Route.Profile> {
             val viewModel = koinViewModel<ProfileViewModel>()
-            LaunchedEffect(Unit) {
-                viewModel.state.collect { event ->
-                    if (event.isSignedOut) {
-                        googleAuthUiClient.signOut()
-                        onSignOut()
-                    }
-                }
-            }
 
             ProfileRoot(
-                onSignOutButton = {
-                    viewModel.onAction(ProfileAction.OnSignOut)
-                },
+                onSignOutButton = { onSignOut() },
                 viewModel = viewModel
             )
         }
@@ -206,39 +158,6 @@ fun HomeNavGraph(
             )
         }
     }
-}
-
-@Composable
-fun RowScope.AddItem(
-    screen: BottomBarScreen,
-    currentDestination: NavDestination?,
-    navController: NavHostController
-) {
-    NavigationBarItem(
-        colors = NavigationBarItemColors(
-            selectedIconColor = MaterialTheme.colorScheme.primary,
-            unselectedIconColor = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.6f),
-            selectedTextColor = MaterialTheme.colorScheme.primary,
-            unselectedTextColor = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.6f),
-            disabledIconColor = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.38f),
-            disabledTextColor = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.38f),
-            selectedIndicatorColor = MaterialTheme.colorScheme.primary.copy(alpha = 0.12f)
-        ),
-        selected = currentDestination?.route == screen.route::class.qualifiedName,
-        onClick = {
-            navController.navigate(screen.route) {
-                popUpTo(navController.graph.findStartDestination().id)
-                launchSingleTop = true
-            }
-        },
-        label = { Text(text = stringResource(screen.title)) },
-        icon = {
-            Icon(
-                imageVector = screen.icon,
-                contentDescription = "${screen.title} icon",
-            )
-        }
-    )
 }
 
 @Composable
